[winforms] imrpove codegeneration
