+++ src/skiasharp/LiveChartsCore.SkiaSharp.WinForms/CartesianChart.cs
@@ -20,35 +20,17 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-// ==============================================================================
-// 
-// this file contains the WinForms specific code for the CartesianChart class,
-// the rest of the code can be found in the _Shared project.
-// 
-// ==============================================================================
-
-using System.Windows.Forms;
-using LiveChartsCore.Drawing;
 using LiveChartsCore.Kernel.Sketches;
-using LiveChartsCore.Measure;
 
 namespace LiveChartsCore.SkiaSharpView.WinForms;
 
-/// <inheritdoc cref="ICartesianChartView" />
-public partial class CartesianChart : ChartControl, ICartesianChartView
-{
-    /// <summary>
-    /// Initializes a new instance of the <see cref="CartesianChart"/> class.
-    /// </summary>
-    public CartesianChart()
-    {
-        GetDrawnControl().MouseWheel += OnMouseWheel;
-    }
+// ==============================================================================
+// 
+// use the LiveChartsGeneratedCode.CartesianChart class to add winforms specific
+// code, this class is just to expose the CartesianChart class in this namespace.
+// 
+// ==============================================================================
 
-    private void OnMouseWheel(object? sender, MouseEventArgs e)
-    {
-        var c = (CartesianChartEngine)CoreChart;
-        var p = e.Location;
-        c.Zoom(ZoomMode, new LvcPoint(p.X, p.Y), e.Delta > 0 ? ZoomDirection.ZoomIn : ZoomDirection.ZoomOut);
-    }
-}
+/// <inheritdoc cref="ICartesianChartView" />
+public partial class CartesianChart : LiveChartsGeneratedCode.CartesianChart
+{ }
