+++ src/skiasharp/LiveChartsCore.SkiaSharp.WinForms/LiveChartsGeneratedCode/PieChart.cs
@@ -22,15 +22,15 @@
 
 using LiveChartsCore.Kernel.Sketches;
 
-namespace LiveChartsCore.SkiaSharpView.WinUI;
+namespace LiveChartsGeneratedCode;
 
 // ==============================================================================
 // 
-// use the LiveChartsGeneratedCode.PieChart class to add winui/uno specific
-// code, this class is just to expose the PieChart class in this namespace.
+// this file contains the WinForms specific code for the PieChart class,
+// the rest of the code can be found in the _Shared project.
 // 
 // ==============================================================================
 
-/// <inheritdoc cref="IChartView" />
-public sealed partial class PieChart : LiveChartsGeneratedCode.PieChart
+/// <inheritdoc cref="IPieChartView" />
+public partial class PieChart : ChartControl, IPieChartView
 { }
