+++ src/skiasharp/LiveChartsCore.SkiaSharp.WinForms/PieChart.cs
@@ -20,17 +20,17 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-// ==============================================================================
-// 
-// this file contains the WinForms specific code for the PieChart class,
-// the rest of the code can be found in the _Shared project.
-// 
-// ==============================================================================
-
 using LiveChartsCore.Kernel.Sketches;
 
 namespace LiveChartsCore.SkiaSharpView.WinForms;
 
+// ==============================================================================
+// 
+// use the LiveChartsGeneratedCode.PieChart class to add winforms specific
+// code, this class is just to expose the PieChart class in this namespace.
+// 
+// ==============================================================================
+
 /// <inheritdoc cref="IPieChartView" />
-public partial class PieChart : ChartControl, IPieChartView
+public partial class PieChart : LiveChartsGeneratedCode.PieChart
 { }
